Refactor TextBufferMarkup to pango_buffer
